Refactor: Split auth into jwt and password modules
